--- ui/app/templates/components/identity/entity-nav.hbs
@@ -1,52 +1,0 @@
-{{!
-  Copyright (c) HashiCorp, Inc.
-  SPDX-License-Identifier: BUSL-1.1
-~}}
-
-<PageHeader as |p|>
-  <p.levelLeft>
-    <h1 class="title is-3">
-      {{titleize (pluralize this.identityType)}}
-    </h1>
-  </p.levelLeft>
-</PageHeader>
-<div class="tabs-container box is-sideless is-fullwidth is-paddingless is-marginless">
-  <nav class="tabs">
-    <ul>
-      <LinkTo @route="vault.cluster.access.identity.index" @model={{pluralize this.identityType}}>
-        {{capitalize (pluralize this.identityType)}}
-      </LinkTo>
-      <LinkTo @route="vault.cluster.access.identity.aliases.index" @model={{pluralize this.identityType}}>
-        Aliases
-      </LinkTo>
-    </ul>
-  </nav>
-</div>
-<Toolbar>
-  {{#if this.model.meta.total}}
-    <ToolbarFilters>
-      <Identity::LookupInput @type={{this.identityType}} />
-    </ToolbarFilters>
-  {{/if}}
-  <ToolbarActions>
-    {{#if (eq this.identityType "entity")}}
-      <ToolbarLink
-        @route={{"vault.cluster.access.identity.merge"}}
-        @model={{pluralize this.identityType}}
-        data-test-entity-merge-link={{true}}
-      >
-        Merge
-        {{pluralize this.identityType}}
-      </ToolbarLink>
-    {{/if}}
-    <ToolbarLink
-      @route="vault.cluster.access.identity.create"
-      @model={{pluralize this.identityType}}
-      @type="add"
-      data-test-entity-create-link={{true}}
-    >
-      Create
-      {{this.identityType}}
-    </ToolbarLink>
-  </ToolbarActions>
-</Toolbar>
